Rename DefaultLogger final
