+++ src/lib/DefaultLogger.js
